Added fetching and parsing data. Added reset_db method. Changed add_const, retrieve_all methods. View now prints data from db.
New method: retrieve_and_store (file) in welcome_controller.rb. Takes a
file address, gets the file and parses it. After that, stores the
content into database.
New method: reset_db(). Deletes all entries in the construction tables.
Changed method: add_constr. Splits the coord string into array now.
Changed ‘In Progress’ to ‘In Process’ to fit the data in the file.
Changed method: retrieve_all. Changed ‘In Progress’ to ‘In Process’ to
fit the data in the file.